Fix types for heft-jest
