+++ types/heft-jest/index.d.ts
@@ -3,10 +3,42 @@
 // Definitions by: Pete Gonzalez <https://github.com/octogonz>
 //                 Ian Clanton-Thuon <https://github.com/iclanton>
 // Definitions: https://github.com/DefinitelyTyped/DefinitelyTyped
-// Minimum TypeScript Version: 4.3
+// Minimum TypeScript Version: 4.5
 
-/// <reference types="jest" />
-/// <reference path="./mocked.d.ts" />
+type MockInstance<T> = import('jest-mock').MockInstance<T>;
+
+namespace mocked {
+    type MockableFunction = (...args: any[]) => any;
+    type MethodKeysOf<T> = { [K in keyof T]: T[K] extends MockableFunction ? K : never }[keyof T];
+    type PropertyKeysOf<T> = { [K in keyof T]: T[K] extends MockableFunction ? never : K }[keyof T];
+    type ArgumentsOf<T> = T extends (...args: infer A) => any ? A : never;
+    type ConstructorArgumentsOf<T> = T extends new (...args: infer A) => any ? A : never;
+
+    interface MockWithArgs<T extends MockableFunction> extends MockInstance<T> {
+        new (...args: ConstructorArgumentsOf<T>): T;
+        (...args: ArgumentsOf<T>): ReturnType<T>;
+    }
+
+    type MaybeMockedConstructor<T> = T extends new (...args: any[]) => infer R
+        ? MockInstance<(...args: ConstructorArgumentsOf<T>) => R>
+        : {};
+    type MockedFunction<T extends MockableFunction> = MockWithArgs<T> & { [K in keyof T]: T[K] };
+    type MockedFunctionDeep<T extends MockableFunction> = MockWithArgs<T> & MockedObjectDeep<T>;
+    type MockedObject<T> = MaybeMockedConstructor<T> &
+        { [K in MethodKeysOf<T>]: T[K] extends MockableFunction ? MockedFunction<T[K]> : T[K] } &
+        { [K in PropertyKeysOf<T>]: T[K] };
+    type MockedObjectDeep<T> = MaybeMockedConstructor<T> &
+        { [K in MethodKeysOf<T>]: T[K] extends MockableFunction ? MockedFunctionDeep<T[K]> : T[K] } &
+        { [K in PropertyKeysOf<T>]: MaybeMockedDeep<T[K]> };
+
+    type MaybeMockedDeep<T> = T extends MockableFunction
+        ? MockedFunctionDeep<T>
+        : T extends object
+        ? MockedObjectDeep<T>
+        : T;
+
+    type MaybeMocked<T> = T extends MockableFunction ? MockedFunction<T> : T extends object ? MockedObject<T> : T;
+}
 
 /**
  * Tests can use the `mocked()` function to access additional properties that
@@ -32,5 +64,5 @@
  *
  * https://kulshekhar.github.io/ts-jest/user/test-helpers
  */
-declare function mocked<T>(item: T, deep?: false): mocked.MaybeMocked<T>;
-declare function mocked<T>(item: T, deep: true): mocked.MaybeMockedDeep<T>;
+function mocked<T>(item: T, deep?: false): mocked.MaybeMocked<T>;
+function mocked<T>(item: T, deep: true): mocked.MaybeMockedDeep<T>;
